pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dd-llm"
version = "0.1.0"
description = "Shared LLM abstraction layer for Digital Duck projects"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Digital Duck", email = "p2p2learn@outlook.com"},
]
keywords = ["llm", "digital-duck", "ai", "provider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.25.0"]
gemini = ["google-genai>=1.0.0"]
all = ["openai>=1.0.0", "anthropic>=0.25.0", "google-genai>=1.0.0"]
dev = ["pytest>=8.0"]

[project.urls]
Homepage   = "https://github.com/digital-duck/dd-llm"
Repository = "https://github.com/digital-duck/dd-llm"
Issues     = "https://github.com/digital-duck/dd-llm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dd_llm*"]
